op-acceptance-tests: split sync_tester tests into multiple packages
