+++ op-acceptance-tests/acceptance-tests.yaml
@@ -85,9 +85,37 @@ gates:
       - package: github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/sync_tester_ext_el
         timeout: 30m
 
-  - id: sync-test-op-node-hfs-ext
-    description: "Sync tests for upgrade scenarios with op-node with external networks via the op-sync-tester"
+  - id: sync-test-op-node-hfs-ext-canyon
+    description: "Upgrade scenarios with op-node with external networks via the op-sync-tester - Canyon"
     tests:
-      - package: github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/sync_tester_hfs_ext
+      - package: github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/sync_tester_hfs_ext_canyon
         timeout: 30m
-        run_all: true
+
+  - id: sync-test-op-node-hfs-ext-ecotone
+    description: "Upgrade scenarios with op-node with external networks via the op-sync-tester - Ecotone"
+      - package: github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/sync_tester_hfs_ext_ecotone
+        timeout: 30m
+
+  - id: sync-test-op-node-hfs-ext-fjord
+    description: "Upgrade scenarios with op-node with external networks via the op-sync-tester - Fjord"
+    tests:
+      - package: github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/sync_tester_hfs_ext_fjord
+        timeout: 30m
+
+  - id: sync-test-op-node-hfs-ext-granite
+    description: "Upgrade scenarios with op-node with external networks via the op-sync-tester - Granite"
+    tests:
+      - package: github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/sync_tester_hfs_ext_granite
+        timeout: 30m
+
+  - id: sync-test-op-node-hfs-ext-holocene
+    description: "Upgrade scenarios with op-node with external networks via the op-sync-tester - Holocene"
+    tests:
+      - package: github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/sync_tester_hfs_ext_holocene
+        timeout: 30m
+
+  - id: sync-test-op-node-hfs-ext-isthmus
+    description: "Upgrade scenarios with op-node with external networks via the op-sync-tester - Isthmus"
+    tests:
+      - package: github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/sync_tester_hfs_ext_isthmus
+        timeout: 30m
