+++ op-acceptance-tests/tests/sync_tester/sync_tester_ext_el/init_test.go
